+++ provider/datasource_rediscloud_active_active_subscription.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"github.com/RedisLabs/rediscloud-go-api/redis"
 	"github.com/RedisLabs/rediscloud-go-api/service/subscriptions"
+	"github.com/RedisLabs/terraform-provider-rediscloud/provider/client"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"strconv"
@@ -146,9 +147,9 @@ func dataSourceRedisCloudActiveActiveSubscription() *schema.Resource {
 
 func dataSourceRedisCloudActiveActiveSubscriptionRead(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
 	var diags diag.Diagnostics
-	api := meta.(*apiClient)
+	api := meta.(*client.ApiClient)
 
-	subs, err := api.client.Subscription.List(ctx)
+	subs, err := api.Client.Subscription.List(ctx)
 	if err != nil {
 		return diag.FromErr(err)
 	}
@@ -214,7 +215,7 @@ func dataSourceRedisCloudActiveActiveSubscriptionRead(ctx context.Context, d *sc
 
 	subId := redis.IntValue(sub.ID)
 
-	m, err := api.client.Maintenance.Get(ctx, subId)
+	m, err := api.Client.Maintenance.Get(ctx, subId)
 	if err != nil {
 		return diag.FromErr(err)
 	}
@@ -222,7 +223,7 @@ func dataSourceRedisCloudActiveActiveSubscriptionRead(ctx context.Context, d *sc
 		return diag.FromErr(err)
 	}
 
-	pricingList, err := api.client.Pricing.List(ctx, subId)
+	pricingList, err := api.Client.Pricing.List(ctx, subId)
 	if err != nil {
 		return diag.FromErr(err)
 	}