+++ provider/rediscloud_active_active_subscription_test.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"flag"
 	"fmt"
+	client2 "github.com/RedisLabs/terraform-provider-rediscloud/provider/client"
 	"regexp"
 	"strconv"
 	"testing"
@@ -111,8 +112,8 @@ func TestAccResourceRedisCloudActiveActiveSubscription_CRUDI(t *testing.T) {
 							return err
 						}
 
-						client := testProvider.Meta().(*apiClient)
-						sub, err := client.client.Subscription.Get(context.TODO(), subId)
+						client := testProvider.Meta().(*client2.ApiClient)
+						sub, err := client.Client.Subscription.Get(context.TODO(), subId)
 						if err != nil {
 							return err
 						}
@@ -315,7 +316,7 @@ func TestAccResourceRedisCloudActiveActiveSubscription_createUpdateMarketplacePa
 }
 
 func testAccCheckActiveActiveSubscriptionDestroy(s *terraform.State) error {
-	client := testProvider.Meta().(*apiClient)
+	client := testProvider.Meta().(*client2.ApiClient)
 
 	for _, r := range s.RootModule().Resources {
 		if r.Type != "rediscloud_active_active_subscription" {
@@ -327,7 +328,7 @@ func testAccCheckActiveActiveSubscriptionDestroy(s *terraform.State) error {
 			return err
 		}
 
-		subs, err := client.client.Subscription.List(context.TODO())
+		subs, err := client.Client.Subscription.List(context.TODO())
 		if err != nil {
 			return err
 		}